Add member to sigstore-codeowners, update rekor-monitor-codeowners
sigstore/sigstore PRs have been sitting without approvals, so I'd like to propose adding myself as a codeowner to unblock development in this repository. I've made a number of contributions to the repository, including some of the code that has recently been moved into the repo recently (fulcio/TUF).

I also updated rekor-monitor-codeowners for two members to not be able to maintain the team, but still be a member.

Signed-off-by: Hayden B <hblauzvern@google.com>