Add RequireLastPushApproval to all repositories (#263)
This will mitigate insider risk where a maintainer can go to any open PR, modify the contents, and approve and merge the PR. This is effectively self-approval and bypasses the minimum of 1 reviewer check, since the maintainer the entire PR without another review. This check will require that the PR approver be different than the author on the last commit. I have set this in every repository except sigstore-blog, since it requires 3 approvals.

See https://github.blog/changelog/2022-10-20-new-branch-protections-last-pusher-and-locked-branch/#require-approval-from-someone-other-than-the-last-pusher for more information.

Note that the only time I have seen this used is when a change to a dependabot PR is needed. This will now require an approver besides the one fixing the dependabot PR.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>