Add Sigstore claimant model
The Sigstore claimant model formally defines a set of claims made by
Sigstore's two transparency logs. This was created in collaboration with
the creators of the claimant model.

This also includes sequence diagrams to make it easier to understand how
the various actors interact. These are a work in progress and generated
by a tool in Trillian.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>