Update sigstore-go maintainers
Adding two new maintainers to sigstore-go in preparation for the update
to the sigstore-go library. Removing inactive members, and we can circle
back to making sure we have a good set of active maintainers in the
coming weeks.

Also removing the sigstore-go-triage group as it's not needed and was
only for the old sigstore library.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>