Add tips regarding teams to prevent last confusion
Last time, when I tried to update the `model_transparency` repo, I added
a `model_transparency` team and added it to the repo in the same PR
(#454), with a plan to fix team ID in the next one (#456). This actually
caused a problem: the `model_transparency_codeowners` team lost access
to the repository.

This was because the team ID was reused, so Pulumi removed permissions
when creating the new team.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>