Properly configure the required checks for model-transparency
I was wrong in #478: the checks need to be as they show up in the job
name, not how they show up in the GHA summary at the bottom of the PR.

Since I used the wrong ones,
https://github.com/sigstore/model-transparency/pull/284 is now waiting
on checks that cannot run. But I cna fix that by closing and reopening
once the required checks are in the correct format.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>